[Update] 재료 등록, 수정, 검색 API 수정

+++ src/main/java/refree/backend/module/Ingredient/IngredientRepository.java
@@ -3,6 +3,6 @@ package refree.backend.module.Ingredient;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IngredientRepository extends JpaRepository<Ingredient, Integer>,IngredientRepositoryCustom {
+public interface IngredientRepository extends JpaRepository<Ingredient, Long>,IngredientRepositoryCustom {
 
 }
